Refactored import statements in Code and Documentation models to streamline module references. Added index.ts to centralize type definitions for BaseRelation, BaseNode, and EmbeddingsContainer, enhancing code organization and maintainability.
